# contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: TUBA Product API
  version: 0.1.0
  description: Current API contract. Operations marked planned are design contracts and are not implemented endpoints.
servers:
  - url: /api/v1
security:
  - oidc: []
paths:
  /ingest/events:
    post:
      operationId: ingestRawEvent
      summary: Accept one source event into the durable raw Kafka stream
      security:
        - sourceApiKey: []
      parameters:
        - name: X-Source-Position
          in: header
          required: true
          description: Stable source position. source-adapter encodes the immutable ingress stream, topic, partition, and offset; producers cannot supply trusted tenant identity.
          schema: {type: string, minLength: 1, maxLength: 512}
        - {name: X-Source-Context, in: header, required: true, schema: {type: string, pattern: '^ctx_[a-f0-9]{32}$'}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: One JSON source event, at most 1 MiB. Beat Kafka input follows the BeatIngressEvent schema; identity and routing metadata are resolved from the registered source context.
              additionalProperties: true
      responses:
        "202":
          description: Raw event durably accepted by Kafka
          content:
            application/json:
              schema:
                type: object
                required: [receipt_id, raw_event_id, source_context_id, source_position, payload_hash, accepted_at, status]
                properties:
                  receipt_id: {type: string}
                  raw_event_id: {type: string}
                  source_context_id: {type: string, pattern: '^ctx_[a-f0-9]{32}$'}
                  source_position: {type: string, minLength: 1, maxLength: 512}
                  payload_hash: {type: string, pattern: '^[a-f0-9]{64}$'}
                  accepted_at: {type: string, format: date-time}
                  status: {type: string, const: accepted}
        "400": {description: Invalid JSON or missing source position}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409": {description: Source position was previously used for a different payload}
        "413": {description: Payload exceeds 1 MiB}
        "415": {description: Content-Type must be application/json}
        "429": {description: Rate limited; Retry-After is set}
        "503": {description: Kafka is temporarily unavailable; retry with the same source position}
  /internal/ingest/beat-events:
    post:
      operationId: ingestBeatEventFromBoundTopic
      summary: Accept a Beat event from a server-bound Kafka source topic
      description: Internal adapter endpoint. Resolves organization, source instance and immutable context from the registered topic; callers cannot provide source identity in the body.
      security: []
      parameters:
        - {name: X-Source-Adapter-Token, in: header, required: true, schema: {type: string, minLength: 32}}
        - {name: X-Source-Topic, in: header, required: true, schema: {type: string, pattern: '^tuba\.source\.ctx_[a-f0-9]{32}\.v1$'}}
        - {name: X-Source-Partition, in: header, required: true, schema: {type: integer, minimum: 0}}
        - {name: X-Source-Offset, in: header, required: true, schema: {type: integer, minimum: 0}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/BeatIngressEvent"}
      responses:
        "202":
          description: Raw event durably accepted; adapter verifies the returned context, topic position and payload hash before committing Kafka offset.
          content:
            application/json:
              schema:
                type: object
                required: [receipt_id, raw_event_id, source_context_id, source_position, payload_hash, accepted_at, status]
                properties:
                  receipt_id: {type: string}
                  raw_event_id: {type: string}
                  source_context_id: {type: string, pattern: '^ctx_[a-f0-9]{32}$'}
                  source_position: {type: string, pattern: '^kafka-v1:tuba\.source\.ctx_[a-f0-9]{32}\.v1:[0-9]+:[0-9]+$'}
                  payload_hash: {type: string, pattern: '^[a-f0-9]{64}$'}
                  accepted_at: {type: string, format: date-time}
                  status: {type: string, const: accepted}
        "400": {description: Invalid source topic position or content type; adapter retains offset}
        "401": {description: Invalid adapter token; adapter retains offset}
        "413": {description: Payload exceeds 1 MiB}
        "429": {description: Rate limited; adapter retains offset and retries}
        "503": {description: Topic binding or downstream dependency unavailable; adapter retains offset and retries}
  /sources:
    get:
      operationId: listSources
      x-required-permission: source:manage
      summary: List source instances in the authenticated tenant
      parameters:
        - {$ref: "#/components/parameters/Cursor"}
        - {$ref: "#/components/parameters/Limit"}
      responses:
        "200":
          description: Source configuration without credentials
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/SourceInstance"}}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      operationId: registerSource
      x-required-permission: source:manage
      summary: Register a source and issue its API key once
      parameters:
        - {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SourceRegistrationInput"}
      responses:
        "201":
          description: Registered source; api_key is only returned at creation
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RegisteredSource"}
        "400": {description: Invalid registration or release is not active/staged}
        "403": {$ref: "#/components/responses/Forbidden"}
        "503": {description: Source registry unavailable}
  /sources/{id}:
    delete:
      operationId: revokeSource
      x-required-permission: source:manage
      summary: Disable a source credential
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, pattern: "^src_[a-f0-9]{32}$"}}
      responses:
        "204": {description: Source disabled}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {description: Source not found}
        "503": {description: Source registry unavailable}
  /sources/{id}/rotate:
    post:
      operationId: rotateSourceCredential
      x-required-permission: source:manage
      summary: Rotate a source key without changing the source epoch
      description: The previous key remains valid for 24 hours for queue draining. Store the returned API key; it is shown only once.
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, pattern: "^src_[a-f0-9]{32}$"}}
      responses:
        "200":
          description: Updated source and newly issued one-time API key
          content:
            application/json:
              schema: {$ref: "#/components/schemas/RegisteredSource"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {description: Source not found or disabled}
        "503": {description: Source registry unavailable}
...
components:
  securitySchemes:
    sourceApiKey:
      type: apiKey
      in: header
      name: X-API-Key
    collectorBearer:
      type: http
      scheme: bearer
      bearerFormat: TUBA Collector credential
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example/realms/tuba/.well-known/openid-configuration
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: {type: string, maxLength: 2048}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
    Forbidden:
      description: Effective permissions do not allow the operation
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
  schemas:
    BeatIngressEvent:
      $ref: ../events/beat-ingress/1/schema.json
...
    SourceRegistrationInput:
      type: object
      additionalProperties: false
      required: [vendor_name, vendor_product, vendor_dataset, release_id, rate_limit]
      properties:
        vendor_name: {type: string, minLength: 1, maxLength: 128}
        vendor_product: {type: string, minLength: 1, maxLength: 128}
        vendor_dataset: {type: string, minLength: 1, maxLength: 128}
        release_id: {type: string, minLength: 1, maxLength: 128}
        rate_limit: {type: integer, minimum: 1, maximum: 100000}
    SourceInstance:
      type: object
      required: [id, organization_id, namespace, vendor_name, vendor_product, vendor_dataset, source_epoch, source_context_id, release_id, enabled, rate_limit, created_at, updated_at]
      properties:
        id: {type: string}
        organization_id: {type: string}
        namespace: {type: string}
        vendor_name: {type: string}
        vendor_product: {type: string}
        vendor_dataset: {type: string}
        source_epoch: {type: string}
        source_context_id: {type: string, pattern: '^ctx_[a-f0-9]{32}$'}
        release_id: {type: string}
        enabled: {type: boolean}
        rate_limit: {type: integer}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    RegisteredSource:
      allOf:
        - {$ref: "#/components/schemas/SourceInstance"}
        - type: object
          required: [api_key]
          properties:
            api_key: {type: string, description: Returned only once; store in source secret storage.}
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code: {type: string}
        message: {type: string}
        request_id: {type: string}